Layout, Favicon, Sequence, Help
Added customizable top buttons to the main layouts.

Changed the favicon to the notes on a staff

Added the sequence generator with both modes and visual checks and
functional report after.

Added read more on the random practicing page with descriptions of how
things all work and what the options are.

Signed-off-by: Rick Henry <rickhenry@rickhenry.dev>